Support nested lists.
Rather than indiscriminately trimming whitespace from key values,
measure the indent on the first line of text and use that as a baseline.

For example, notice that the first line of text after a summary: keyword is indented by
two tabs, and then if three-tab text is encountered, we know it's a nested list or
a code block.